Created python programs for batch processing in Slicer
Change_labels was created to allow for the labels on as many volumes as are imported to be changed. It goes through each volume and changes the specified labels from what they were to a new label, deleting all other labels.

labelStatistics was created to allow for batch processing of mean and other values of a gray image to be calcualted over a segmented ROI using Python. At this first stage the program can do the calcualtion for one volume. The beginning code from Change_labels needs to be implemeneted in this program to allo this to be done in "batch" processing.